+++ heartbeat-distributor-service/src/main/java/edu/gemini/aspen/heartbeatdistributor/IHeartbeatConsumer.java
@@ -12,5 +12,5 @@ public interface IHeartbeatConsumer {
      *
      * @param beatNumber the last heartbeat number received
      */
-    void beat(long beatNumber);
+    void beat(int beatNumber);
 }
